+++ R/get_rcode_header_osprey.R
@@ -1,8 +1,0 @@
-#' Used to get the header for "Show R-Code" - temp solution for osprey/teal.osprey until teal is updated
-#'
-#' @noRd
-#'
-get_rcode_header_osprey <- function(title, datanames, datasets, code_data_processing = NULL) {
-  warning("This function is deprecated")
-  return("")
-}