+++ R/utils.R
@@ -13,11 +13,10 @@
 #' @template author_zhanc107
 #'
 
-quick_filter <- function(filter_opt, ANL){
+quick_filter <- function(filter_opt, ANL) { # nolint
   for (i in seq(1:length(filter_opt))) {
-    ANL <- ANL[ANL[, filter_opt[i]] == "Y", ]
+    ANL <- ANL[ANL[, filter_opt[i]] == "Y", ] # nolint
   }
-
   return(ANL)
 }